requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
tqdm>=4.50.0

# Optional dependencies for enhanced functionality
scikit-image>=0.17.0
scipy>=1.5.0

# Development dependencies (install with pip install -e ".[dev]")
pytest>=6.0.0
pytest-cov>=2.10.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.910
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Example notebook dependencies (install with pip install -e ".[examples]")
jupyter>=1.0.0
ipywidgets>=7.6.0
seaborn>=0.11.0